Struct wasmer_api::global_id::GlobalId

pub struct GlobalId {
    kind: NodeKind,
    database_id: u64,
}

Global id of backend nodes.

IDs are encoded using the “hashid” scheme, which uses a given alphabet and a salt to encode u64 numbers into a string hash.

Fields
kind: NodeKind

The node type of the ID.

database_id: u64

The database ID of the node.

Implementations

impl GlobalId

const SALT: &'static str = "wasmer salt hashid"

Salt used by the backend to encode hashes.

const MIN_LENGTH: usize = 12usize

Minimum length of the encoded hashes.

const ALPHABET_PREFIXED: &'static str = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890"

Hash alphabet used for the prefix id variant.

const ALPHABET_URL: &'static str = "abcdefghijklmnopqrstuvwxyz0123456789"

Hash alphabet used for the non-prefixed id variant.

pub fn new(kind: NodeKind, database_id: u64) -> Self

fn build_harsh(alphabet: &str, salt: &[u8]) -> Harsh

fn build_harsh_prefixed() -> Harsh

fn build_harsh_url() -> Harsh

pub fn kind(&self) -> NodeKind

pub fn database_id(&self) -> u64

pub fn encode_prefixed(&self) -> String

Encode a prefixed global id.

fn parse_values(values: &[u64]) -> Result<Self, ErrorKind>

pub fn parse_prefixed(hash: &str) -> Result<Self, GlobalIdParseError>

Parse a prefixed global id.

pub fn encode_url(&self) -> String

Encode a non-prefixed global id.

Note: URL ids use a different alphabet than prefixed ids.

pub fn parse_url(hash: &str) -> Result<Self, GlobalIdParseError>

Parse a non-prefixed URL global id variant.

Note: URL ids use a different alphabet than prefixed ids.