Module wasmer_api::query

Enums

GenerateTokenBy

TokenKind

Functions

all_app_versions

Load all versions of an app.

app_version_activate

Activate a particular version of an app.

create_namespace

Create a new namespace.

current_user_with_namespaces

Get the currently logged in used, together with all accessible namespaces.

delete_app

Delete an app.

fetch_webc_package

Load a webc package from the registry.

generate_deploy_config_token_raw

generate_deploy_token_raw

Generate a new Edge token.

get_all_dns_records

Retrieve all DNS records.

get_all_dns_records_stream

Retrieve a domain by its name.

get_all_domains

Retrieve all DNS domains.

get_app

Retrieve an app.

get_app_and_package_by_name

Retrieve an app together with a specific version.

get_app_by_alias

Retrieve an app by its global alias.

get_app_by_id

Retrieve an app by its global id.

get_app_by_id_opt

Retrieve an app by its global id.

get_app_logs 🔒

Get pages of logs associated with an application that lie within the specified date range.

get_app_logs_paginated

Get pages of logs associated with an application that lie within the specified date range.

get_app_version

Retrieve an app version.

get_app_version_by_id

Retrieve an app version by its global id.

get_app_version_by_id_with_app

get_app_with_version

Retrieve an app together with a specific version.

get_app_with_version_by_id

Retrieve an app together with a specific version.

get_deploy_app_versions

Retrieve versions for an app.

get_deploy_apps

Retrieve apps.

get_deploy_apps_stream

Retrieve apps as a stream that will automatically paginate.

get_domain

Retrieve a domain by its name.

get_domain_with_records

Retrieve a domain by its name, along with all it’s records.

get_domain_zone_file

Retrieve a domain by its name.

get_namespace

Retrieve a namespace by its name.

get_node

Retrieve a node based on its global id.

get_package

Retrieve a package by its name.

get_package_version

Retrieve a package version by its name.

get_package_versions

Retrieve package versions for an app.

get_package_versions_stream

Retrieve all versions of a package as a stream that auto-paginates.

namespace_apps

Get apps for a specific namespace.

publish_deploy_app

Publish a new app (version).

register_domain

Register a new domain

unix_timestamp 🔒

Convert a [OffsetDateTime] to a unix timestamp that the WAPM backend understands.

upsert_domain_from_zone_file

Publish a new app (version).

user_accessible_apps

List all apps that are accessible by the current user.

user_apps

List all apps that are accessible by the current user.

user_namespaces

Get all namespaces accessible by the current user.