Type Alias wasmer_api::types::schema::ID

pub type ID = Id;

Aliased Type

struct ID(String);

Fields
0: String
Implementations

impl Id

pub fn new(s: impl Into<String>) -> Id

Constructs an ID from a String, &str or similar

cynic::Id::new("123");

pub fn inner(&self) -> &str

Returns a reference to the value of this Id

pub fn into_inner(self) -> String

Converts this Id into its inner value

pub fn from_ref(s: &String) -> &Id

Converts a reference to a String to a reference to an Id

To be used when you can access an &String which you want to assume is an Id for use in Cynic structures without reallocating

If you don’t have a String at hand but only an &str, you should know that these can be used directly in InputObjects as well when the target GraphQL type is an Id.